Bugfix: -edit configuration app laravel -edit coordinate type column to projects and images

+++ laravel/database/migrations/2023_05_18_145906_create_boundary_points_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->foreign('satellite_image_id')->references('id')->on('satellite_images');
 
             $table->integer('position');
-            $table->float('x');
-            $table->float('y');
+            $table->bigInteger('x');
+            $table->bigInteger('y');
             
             $table->timestamps();
         });
